requirements.txt
pandas
snowflake-connector-python
snowflake-connector-python[pandas]
requests
python-dotenv